+++ api/src/services/mail.service.ts
@@ -3,6 +3,7 @@ import { createTransport, getTestMessageUrl, SendMailOptions, Transporter } from
 import { config } from '../config';
 import { ProjectUser } from '../entity/project-user.entity';
 import { User } from '../entity/user.entity';
+import { Invite } from '../entity/invite.entity';
 import * as Handlebars from 'handlebars';
 import { join } from 'path';
 import { fstat, readFileSync } from 'fs';
@@ -17,6 +18,7 @@ export default class MailService {
     passwordChanged: this.mustCompileTemplate('password-changed.txt'),
     passwordResetToken: this.mustCompileTemplate('password-reset-token.txt'),
     invitedToProject: this.mustCompileTemplate('invited-to-project.txt'),
+    invitedToPlatform: this.mustCompileTemplate('invited-to-platform.txt'),
   };
 
   constructor() {
@@ -77,6 +79,18 @@ export default class MailService {
       subject: 'You have been granted access to a project on Traduora',
       text: this.knownTemplates.invitedToProject({
         user: user,
+        virtualHost: config.virtualHost,
+      }),
+    });
+  }
+
+  async invitedToPlatform(invite: Invite) {
+    await this.send({
+      from: this.from,
+      to: invite.email,
+      subject: 'You have been invited to a project on Traduora',
+      text: this.knownTemplates.invitedToPlatform({
+        invite: invite,
         virtualHost: config.virtualHost,
       }),
     });