+++ api/test/user.e2e-spec.ts
@@ -1,3 +1,4 @@
+import { Connection } from 'typeorm';
 import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { getRepository } from 'typeorm';
@@ -167,7 +168,7 @@ describe('UserController (e2e)', () => {
   it('/api/v1/users/me (DELETE) should not delete a users account if he is last project user and project has a team', async () => {
     // Add project user (non-admin)
     await request(app.getHttpServer())
-      .post(`/api/v1/projects/${testProject.id}/users`)
+      .post(`/api/v1/projects/${testProject.id}/invites`)
       .set('Authorization', `Bearer ${testingUser3.accessToken}`)
       .send({
         email: testingUser2.email,
@@ -268,6 +269,7 @@ describe('UserController (e2e)', () => {
   });
 
   afterEach(async () => {
+    await app.get(Connection).close();
     await app.close();
   });
 });