+++ webapp/src/app/projects/models/project-invite.ts
@@ -1,0 +1,7 @@
+import { ProjectRole } from './project-role';
+
+export class ProjectInvite {
+  id: string;
+  email: string;
+  role: ProjectRole;
+}